Revert "u8815: switch to 109808 RIL"
Reports of a lot of people not being able to use network.
I can't have this breaking RIL for more people than it solves.
The minority will have to suck it up and find a way to get their
odd networks to work on 2030 RIL on 109808. It didn't even work
with this anyway.

This reverts commit ac1a097c56a9d9e5858f45c2a1de24534679081d.

Conflicts:
	u8815/proprietary/bin/bridgemgrd
	u8815/proprietary/bin/port-bridge
	u8815/proprietary/bin/qmiproxy